fix Solaris compilation bug (#4705)
git-svn-id: https://svn.osgeo.org/gdal/branches/1.9@24594 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/netcdf/netcdfdataset.cpp
@@ -964,7 +964,7 @@ void  netCDFRasterBand::CheckValidData ( void * pImage, int bCheckIsNan )
             /* check for nodata and nan */
             if ( CPLIsEqual( (double) ((T *)pImage)[i], dfNoDataValue ) )
                 continue;
-            if( bCheckIsNan && CPLIsNan( ( (T *) pImage)[i] ) ) {
+            if( bCheckIsNan && CPLIsNan( (double) (( (T *) pImage))[i] ) ) { 
                 ( (T *)pImage )[i] = (T)dfNoDataValue;
                 continue;
             }
